move_h5_data changes
Finalised comments to move_h5_data.py file

+++ prepare_data/move_h5_data.py
@@ -21,12 +21,13 @@ TODO: Need to change the way I read the groups for efficiency
 '''
 def get_curent_ratio(h5_file):
     total = 0
+    # read the h5 file and get the number of slices
     with h5py.File(h5_file, "r") as hf:
         curr_test = np.array(hf["/test/patients"]).size
         curr_train = np.array(hf["/train/patients"]).size
         curr_val = np.array(hf["/validation/patients"]).size
         
-    total = curr_test + curr_train + curr_val
+    total = curr_test + curr_train + curr_val   #add slices from each group to get total
 
     return curr_test, curr_train, curr_val, total
 
@@ -71,6 +72,10 @@ def calculate_differences(curr_test, curr_train, curr_val, trgt_test, trgt_train
 
 '''
 This function calculates the number of slices for each group
+Input:  h5_file = path to h5 file
+        group = current group
+
+Output: slice_count = slices remaining
 '''
 def get_slice_count(h5_file, group):
     with h5py.File(h5_file, "r") as hf:
@@ -80,6 +85,11 @@ def get_slice_count(h5_file, group):
 
 '''
 This function adds the data needed to the model
+Input:  dsm = DataSetModel
+        cine_grp = the cine group from h5 file
+        tagged_grp = the tagged group from h5 file
+        start = index of first slice
+        end = index of last slice
 '''
 def add_data_to_model(dsm, grp, cine_grp, tagged_grp, start, end):
     dsm.patient_names = np.array(grp.get("patients")[start:end])
@@ -89,7 +99,7 @@ def add_data_to_model(dsm, grp, cine_grp, tagged_grp, start, end):
     dsm.cim_paths = np.array(grp.get("cim_paths")[start:end])
 
     # cines
-    dsm.cine_dicom_paths = np.array(cine_grp.get("cine_dicom_paths")[start:end, :])
+    dsm.cine_dicom_paths = np.array(cine_grp.get("cine_dicom_paths")[start:end,:])
     dsm.cine_img_orient = np.array(cine_grp.get("cine_image_orientations")[start:end,:])
     dsm.cine_img_pos = np.array(cine_grp.get("cine_image_positions")[start:end,:])
     dsm.cine_images = np.array(cine_grp.get("cine_images")[start:end,:,:,:])
@@ -103,7 +113,15 @@ def add_data_to_model(dsm, grp, cine_grp, tagged_grp, start, end):
     dsm.tagged_px_spaces = np.array(tagged_grp.get("tagged_px_spaces")[start:end])
 
 '''
+This function takes data from the current group. It gets the slices from the patient at the end
+and returns the index number of the first slice of that patient (from the right, as a positive number).
+Inputs: dsm = datasetmodel
+        h5_file = path to the h5 file we;re taking data from
+        group = group we're taking data from
+        index = index of the first slice of the last patient taken (from the right)
+        slice_count = current number of slices in that group
 
+Output: index = index of the first slice of the patient taken (from the right)
 '''
 def get_data_from_group(dsm, h5_file, group, index, slice_count):
     with h5py.File(h5_file, "r") as hf:
@@ -114,7 +132,8 @@ def get_data_from_group(dsm, h5_file, group, index, slice_count):
 
         # get the list of patients
         try:
-            patients = np.array(grp.get("patients")[slice_count-5:slice_count])   #so we don't have to read the entire patient array
+            #so we don't have to read the entire patient array, slices range from 1-3 so taking 5 patients is good
+            patients = np.array(grp.get("patients")[slice_count-5:slice_count])   
         except IndexError:
             patients = np.array(grp.get("patients")[:slice_count])
         patient = patients[-1] #take only one patient
@@ -162,7 +181,7 @@ def get_data_for_group(dsm, h5_file, group, start, end):
     
 '''
 This function copies h5 data from old h5 file to the new h5 file
-Input: h5_file = path to the old h5 file
+Input:  h5_file = path to the old h5 file
         new_h5_file = path to the new h5 file
         group = group we're copying
         init = whether the h5 file is already created or not
@@ -182,7 +201,7 @@ def copy_h5_data(h5_file, new_h5_file, group, init, limit):
     
     slice_cnt = get_slice_count(h5_file, group) #get the slice count
     if limit is not None:   #if the last index(from the right) is specified
-        slice_cnt = slice_cnt - (limit-1)
+        slice_cnt = slice_cnt - (limit-1) 
     for i in range(0,slice_cnt,5):    #loop through the data in small increments to not overload the memory(if we have huge h5 files)
         dsm = DataSetModel()
         if (slice_cnt - i) < 5:   #if we've reached the end of the data
@@ -207,6 +226,8 @@ Overview:
     1. Calculates the difference between the target number of slices and the current number of slices for each group
     2. Moves data from groups (old h5 file) that have number of slices that are more than desired number of slices to 
     groups (new h5 file) that have number of slices less than the desired number of slices.
+    -The movement of slices per iteration is patient based. (i.e. we move all slices for one patient from the old h5 file
+    to the new h5 file)
 
 How data is moved/copied:
     There are three cases:
@@ -214,8 +235,7 @@ How data is moved/copied:
     2. Adding data to 2 groups from 1 group
     3. Adding data from 1 group to 1 group (one group unchanged)
 
-    First case(Taking data from 2 groups and adding to 1 group):
-    1. Copy the group we're adding to from the old h5 file to the new h5 file
+    1. Copy the group/s we're adding to from the old h5 file to the new h5 file
     2. Loop through the groups we're taking data from
         -Get all the data for all slices from a single patient(starting from the end)
         -Track the index of the first slice of that patient (this becomes our new index for the group we're taking data from)
@@ -226,8 +246,15 @@ How data is moved/copied:
         -Move on to the next group
 
     Second case(Adding data to 2 groups from 1 group):
-    Difference from the first case: Copy from both of the groups we're adding data to to the new h5 file
+    Difference from the first case: 
+    -copies from both of the groups we're adding data to to the new h5 file
+    -loops through the groups we're adding data to instead of group we're taking data from
     -Once all target number of slices met, we copy data from the group we're getting data from (only up until the end index)
+
+    Third case (Adding data to 1 group from one group):
+    Difference from the first case:
+    -copies the data from unchanged group, and group we're adding to to the new h5 file
+    -no loops
 '''
 def move_h5_data(h5_file, new_h5_file, ratio):
     # groups in the h5 file
@@ -257,7 +284,8 @@ def move_h5_data(h5_file, new_h5_file, ratio):
             grp_from.append(grp)
 
     # start moving data
-    if len(grp_from) > len(grp_to): #if we're moving data from 2 groups to 1 group
+    # First case -> if we're moving data from 2 groups to 1 group
+    if len(grp_from) > len(grp_to): 
         print("Moving data from {} & {} to {}".format(grp_from[0], grp_from[1], grp_to[0]))
         init = True #true means the h5 file hasn't been created yet
         # copy the data in the receiving group from the old h5 file to the new h5 file
@@ -266,7 +294,8 @@ def move_h5_data(h5_file, new_h5_file, ratio):
         # loop through the groups that we're taking data from
         for grp in grp_from:
             print("From {} to {}".format(grp, grp_to[0]))
-            #index keeps track of the number of data we're copying from the from group in the old h5 file to the group in the new h5 file
+            # index keeps track of the last index of slice we're copying from the group we're taking data from 
+            # in the old h5 file to the group in the new h5 file
             index = 1   
             slice_cnt = get_slice_count(h5_file, grp)   #get the current number of slices
             # loop until the number of slices is equal to or less than the target number of slices for that group
@@ -278,54 +307,79 @@ def move_h5_data(h5_file, new_h5_file, ratio):
                 slice_cnt -= len(dsm.slices)    #update the number of slices of the group we're taking data from
             copy_h5_data(h5_file, new_h5_file, grp, init, index)    #if target number of slices is reached, we copy the data from the current group to the new h5 file
     
-    elif len(grp_to) > len(grp_from):   #if we're moving data to 2 groups from 1 group
+    # Second case ->  if we're moving data to 2 groups from 1 group
+    elif len(grp_to) > len(grp_from):  
         print("Moving data to {} & {} from {}".format(grp_to[0], grp_to[1], grp_from[0]))
         init = True #true means the h5 file hasn't been created yet
+
         # copy the data in the receiving groups from the old h5 file to the new h5 file
         copy_h5_data(h5_file, new_h5_file, grp_to[0], init, None)
         init = False
         copy_h5_data(h5_file, new_h5_file, grp_to[1], init, None)
-        #index keeps track of the number of data we're copying from the from group in the old h5 file to the group in the new h5 file
+
+        # index keeps track of the last index of slice we're copying from the group we're taking data from 
+        # in the old h5 file to the group in the new h5 file
         index = 1
-        slice_cnt_from = get_slice_count(h5_file, grp_from[0])   #get the current number of slices
+        slice_cnt_from = get_slice_count(h5_file, grp_from[0])   #get the current number of slices of group we're taking data from
         # loop through the groups we're adding data to
         for grp in grp_to:
             print("From {} to {}".format(grp_from[0], grp))
-            slice_cnt = get_slice_count(h5_file, grp)   #get the current number of slices
-            # loop until we get the desired number of slices
+            slice_cnt = get_slice_count(h5_file, grp)   #get the current number of slices of the group we're adding to
+            # loop until we get the desired number of slices for the receiving group
             while slice_cnt < trgt_dict[grp]:   #if the number of slices for the current group reaches the target, we move on to the next group
                 dsm = DataSetModel()     #initiate the model
-                index = get_data_from_group(dsm, h5_file, grp_from[0], index, slice_cnt_from)
+                index = get_data_from_group(dsm, h5_file, grp_from[0], index, slice_cnt_from)   #get data from group we're getting data from
                 print("Moving {} slices for patient {}".format(len(dsm.slices), dsm.patient_names[0]))  
                 add_data_to_group(dsm, new_h5_file, grp)    #add data to new h5 file
-                slice_cnt += len(dsm.slices)    #update the number of slices
+                # update the number of slices for both receiving, and giving groups
+                slice_cnt += len(dsm.slices)
                 slice_cnt_from -= len(dsm.slices)
         #if target number of slices is reached, we copy the data from the current group to the new h5 file
         copy_h5_data(h5_file, new_h5_file, grp_from[0], init, index)    
 
+    # Third case -> if we're moving data from 1 group to another
     else:   #if one of the groups contains the target amount of data
         print("Moving data from {} to {}".format(grp_from[0], grp_to[0]))
         # get the group that already reached its target number of slices
         unchanged_grp = [g for g in groups if g not in grp_to and g not in grp_from]
         print("{} group unchanged".format(unchanged_grp[0]))
-        init = True #true measn the h5 file hasn't been created yet
+
+        # copying data
+        init = True #true means the h5 file hasn't been created yet
         copy_h5_data(h5_file, new_h5_file, unchanged_grp[0], init, None)    #just copy the unchanged group from old h5 file to new h5 file
         init = False
         copy_h5_data(h5_file, new_h5_file, grp_to[0], init, None)   #copy the group we're adding data to
+
+        # index keeps track of the last index of slice we're copying from the group we're taking data from 
+        # in the old h5 file to the group in the new h5 file
         index = 1
-        slice_cnt = get_slice_count(h5_file, grp_from[0])
-        while slice_cnt > trgt_dict[grp_from[0]]:
-            dsm = DataSetModel()
+        slice_cnt = get_slice_count(h5_file, grp_from[0])   #get the current number of slices in the giving group
+        while slice_cnt > trgt_dict[grp_from[0]]:   #move slices until we reach the target number of slices
+            dsm = DataSetModel()    #initiate the model
             index = get_data_from_group(dsm, h5_file, grp_from[0], index, slice_cnt)
             print("Moving {} slices for patient {}".format(len(dsm.slices), dsm.patient_names[0]))
-            add_data_to_group(dsm, new_h5_file, grp_to[0])
-            slice_cnt -= len(dsm.slices)
+            add_data_to_group(dsm, new_h5_file, grp_to[0])  #add data to new h5 file
+            slice_cnt -= len(dsm.slices)    #update number of slices
         copy_h5_data(h5_file, new_h5_file, grp_from[0], init, index)
 
+
+    # get new ratio/division
     curr_test, curr_train, curr_val, total = get_curent_ratio(new_h5_file)
     print("New data distribution -> Test: {} Train: {} Validation: {}".format(curr_test, curr_train, curr_val))
     print("Total: ", total)
 
+'''
+Things to modify:
+input_dir = where the old h5 file you want to modify is located
+h5_filename = filename of you h5 file
+
+output_dir = where you want the new h5 file to be stored
+new_h5_filename = what you want the new h5 file to be called (must be a different name from the old one)
+
+ratio = how you want your data to be divided (must be equal to 1) [test, train, validation]
+
+del_old = if you want to delete the old h5 file
+'''
 if __name__ == "__main__":
     # specify where the h5 file is located
     input_dir = "C:\\Users\\arad572\\Documents\\Summer Research\\code\\prepare_data\\h5_files"
@@ -336,22 +390,28 @@ if __name__ == "__main__":
     output_dir = "C:\\Users\\arad572\\Documents\\Summer Research\\code\\prepare_data\\h5_files"
     new_h5_filename = "UK_Biobank_50cases_new.h5"   #specify the new filename
     new_h5_file = os.path.join(output_dir, new_h5_filename)
-
+        
     # specify how you want to divide the data [test, train, validation], total must equal to 1
     ratio = [0.1, 0.8, 0.1]
 
     # change to False if you don't want to keep the old h5 file
     del_old = False
-    print("Moving h5 data from {} to {}".format(os.path.basename(h5_file), os.path.basename(new_h5_file)))
-    print("h5 files in {}".format(os.path.dirname(h5_file)))
 
-    move_h5_data(h5_file, new_h5_file, ratio)   #move data
+    if h5_file == new_h5_file:
+        print("Please specify a different output filename/path")
+    elif sum(ratio) != 1:
+        print("Sum of ratios must equal 1")
+    else:
+        print("Moving h5 data from {} to {}".format(os.path.basename(h5_file), os.path.basename(new_h5_file)))
+        print("h5 files in {}".format(os.path.dirname(h5_file)))
 
-    if del_old: #if we don't want to keep the old file(saves space)
-        print("Deleting {}".format(h5_filename))
-        os.remove(h5_file)
-        print("Renaming {} to {}".format(new_h5_filename, h5_filename))
-        os.rename(os.path.join(output_dir, new_h5_filename), os.path.join(output_dir, h5_filename))
-        print("New h5 file stored in: {}".format(os.path.join(output_dir, h5_file)))
-    
-    print("Finished")
+        move_h5_data(h5_file, new_h5_file, ratio)   #move data
+
+        if del_old: #if we don't want to keep the old file(saves space)
+            print("Deleting {}".format(h5_filename))
+            os.remove(h5_file)
+            print("Renaming {} to {}".format(new_h5_filename, h5_filename))
+            os.rename(os.path.join(output_dir, new_h5_filename), os.path.join(output_dir, h5_filename)) #renaming to old filename
+            print("New h5 file stored in: {}".format(os.path.join(output_dir, h5_file)))
+        
+        print("Finished")
